fix preload
