+++ src/components/index.jsx
@@ -9,7 +9,7 @@ import Bio from "./Bio/Bio";
 import Contact from "./Contact/Contact";
 import Draw from "./Draw";
 import ModalStain from "./ModalStain";
-import SEO from "./SEO/SEO";
+import SEO from "./SEO";
 
 import useVisualMode from "../hooks/useVisualMode";
